+++ src/API/controller/create/createRoutes.js
@@ -6,12 +6,7 @@ const schema = require('../../middleware/joiSchema.js');
 
 const createController = require('./createController.js');
 
-router.post('/create', validation.validate(schema.book), createController.create);
-router.post('/createAuthor', createController.createAuthor);
-router.post('/createLanguage', createController.createLanguage);
-router.post('/createBook', createController.createBook);
-
-router.post('/insertBooks', createController.insertBooks);
+router.post('/insertBooks', validation.validate(schema.insert) , createController.insertBooks);
 
 module.exports = router;
